improved my_datetime
As mentioned in the PR for added random tests for my_datetime, so far there's something causing the last day of a leap year returning the wrong year. For example: my_datetime(220838400) should be 12-31-1976 but we get 12-31-1977 instead. 
I haven't figured out a way to fix this while avoiding causing more issues at the same time.